+++ TabletopTweaks/Config/SettingGroup.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 
 namespace TabletopTweaks.Config {
-    public class SettingGroup {
+    public class SettingGroup : ICollapseableGroup {
         public bool DisableAll = false;
         public SortedDictionary<string, SettingData> Settings = new SortedDictionary<string, SettingData>();
         public virtual bool this[string key] => IsEnabled(key);
+        public bool IsExpanded = true;
 
         public void LoadSettingGroup(SettingGroup group, bool frozen) {
             DisableAll = group.DisableAll;
@@ -31,6 +32,22 @@ namespace TabletopTweaks.Config {
         public virtual bool IsDisabled(string key) {
             return !IsEnabled(key);
         }
+
+        public void ChangeSetting(string key, bool value) {
+            if (this.DisableAll) {
+                return;
+            }
+            Settings[key].Enabled = value;
+        }
+
+        ref bool ICollapseableGroup.IsExpanded() {
+            return ref IsExpanded;
+        }
+
+        public void SetExpanded(bool value) {
+            IsExpanded = value;
+        }
+
         public class SettingData {
             public bool Enabled;
             public bool Homebrew;